Make CardGameManager now a HeistManager
